document + move Possibilities to a new file
