+++ explore/lower-your-guards/lower-your-guards.cabal
@@ -45,4 +45,5 @@ executable lower-your-guards
                        Play,
                        Annotated,
                        MatchTree,
-                       UA
+                       UA,
+                       Possibilities
